+++ CHANGELOG.md
@@ -22,5 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Infrastructure
 
+- Build System: Refactored plugin deployment
 - CI: Synchronize labels from config file
 - CI: Publish binary packages and debian packages